+++ nym-vpn-core/nym-vpn-lib/src/mixnet/connect.rs
@@ -2,15 +2,10 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 use nym_config::defaults::NymNetworkDetails;
-use nym_sdk::mixnet::{
-    MixnetClient, MixnetClientBuilder, MixnetClientSender, NodeIdentity, Recipient, StoragePaths,
-};
-#[cfg(target_family = "unix")]
-use std::os::fd::RawFd;
+use nym_sdk::mixnet::{MixnetClientBuilder, NodeIdentity, StoragePaths};
 #[cfg(not(target_family = "unix"))]
 use std::os::raw::c_int as RawFd;
 use std::path::PathBuf;
-use std::sync::Arc;
 use tracing::{debug, info};
 
 use crate::{
@@ -19,45 +14,7 @@ use crate::{
     MixnetClientConfig,
 };
 
-#[derive(Clone)]
-pub struct SharedMixnetClient(Arc<tokio::sync::Mutex<Option<MixnetClient>>>);
-
-impl SharedMixnetClient {
-    pub fn new(mixnet_client: MixnetClient) -> Self {
-        Self(Arc::new(tokio::sync::Mutex::new(Some(mixnet_client))))
-    }
-
-    pub async fn lock(&self) -> tokio::sync::MutexGuard<'_, Option<MixnetClient>> {
-        self.0.lock().await
-    }
-
-    pub async fn nym_address(&self) -> Recipient {
-        *self.lock().await.as_ref().unwrap().nym_address()
-    }
-
-    pub async fn split_sender(&self) -> MixnetClientSender {
-        self.lock().await.as_ref().unwrap().split_sender()
-    }
-
-    pub async fn gateway_ws_fd(&self) -> Option<RawFd> {
-        self.lock()
-            .await
-            .as_ref()
-            .unwrap()
-            .gateway_connection()
-            .gateway_ws_fd
-    }
-
-    pub async fn disconnect(self) -> Self {
-        let handle = self.lock().await.take().unwrap();
-        handle.disconnect().await;
-        self
-    }
-
-    pub fn inner(&self) -> Arc<tokio::sync::Mutex<Option<MixnetClient>>> {
-        self.0.clone()
-    }
-}
+use super::SharedMixnetClient;
 
 fn true_to_enabled(val: bool) -> &'static str {
     if val {
