+++ nym-vpn-core/nym-vpn-lib/src/mixnet/mod.rs
@@ -1,0 +1,11 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+mod connect;
+mod mixnet_listener;
+mod processor;
+mod shared_mixnet_client;
+
+pub(crate) use connect::setup_mixnet_client;
+pub(crate) use processor::{start_processor, Config};
+pub(crate) use shared_mixnet_client::SharedMixnetClient;
