+++ nym-vpn-core/nym-vpn-lib/src/mixnet/shared_mixnet_client.rs
@@ -1,0 +1,49 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use nym_sdk::mixnet::{MixnetClient, MixnetClientSender, Recipient};
+#[cfg(target_family = "unix")]
+use std::os::fd::RawFd;
+#[cfg(not(target_family = "unix"))]
+use std::os::raw::c_int as RawFd;
+use std::sync::Arc;
+
+#[derive(Clone)]
+pub struct SharedMixnetClient(Arc<tokio::sync::Mutex<Option<MixnetClient>>>);
+
+impl SharedMixnetClient {
+    pub fn new(mixnet_client: MixnetClient) -> Self {
+        Self(Arc::new(tokio::sync::Mutex::new(Some(mixnet_client))))
+    }
+
+    pub async fn lock(&self) -> tokio::sync::MutexGuard<'_, Option<MixnetClient>> {
+        self.0.lock().await
+    }
+
+    pub async fn nym_address(&self) -> Recipient {
+        *self.lock().await.as_ref().unwrap().nym_address()
+    }
+
+    pub async fn split_sender(&self) -> MixnetClientSender {
+        self.lock().await.as_ref().unwrap().split_sender()
+    }
+
+    pub async fn gateway_ws_fd(&self) -> Option<RawFd> {
+        self.lock()
+            .await
+            .as_ref()
+            .unwrap()
+            .gateway_connection()
+            .gateway_ws_fd
+    }
+
+    pub async fn disconnect(self) -> Self {
+        let handle = self.lock().await.take().unwrap();
+        handle.disconnect().await;
+        self
+    }
+
+    pub fn inner(&self) -> Arc<tokio::sync::Mutex<Option<MixnetClient>>> {
+        self.0.clone()
+    }
+}
